requirements.txt
wireviz-web
flask
pywebview
requests